+++ src/core/config/index.ts
@@ -1,3 +1,3 @@
 export * from "./error";
 export * from "./app";
-export * from "./types";
+export * from "./utils";