+++ src/core/config/types.ts
@@ -1,5 +1,0 @@
-export type ApiResponse<Data> = {
-  data: Data;
-  errors?: string[] | null;
-  status?: number;
-};