+++ src/core/config/utils.ts
@@ -1,0 +1,15 @@
+export type ApiResponse<Data> =
+  | {
+      data: Data;
+      errors?: null;
+      status?: number;
+    }
+  | {
+      data: null;
+      errors: string[];
+      status?: number;
+    };
+
+export const isApiResponse = <Data>(data: any): data is ApiResponse<Data> => {
+  return "data" in data && "errors" in data;
+};