+++ src/users/useCases/activateUser.ts
@@ -1,6 +1,6 @@
 import ms from "ms";
 import { ERROR_CODE, handleErrors } from "../../core/config/error";
-import { ApiResponse } from "../../core/config/types";
+import { ApiResponse } from "../../core/config/utils";
 import { DeleteManyByIdRepository } from "../../core/data/repositories";
 import {
   EmailTemplate,