+++ src/users/useCases/resetPassword.ts
@@ -1,5 +1,5 @@
 import { ERROR_CODE, handleErrors } from "../../core/config/error";
-import { ApiResponse } from "../../core/config/types";
+import { ApiResponse } from "../../core/config/utils";
 import {
   EmailTemplate,
   MailerRepository,