+++ src/users/util/index.ts
@@ -1,3 +1,3 @@
-export * as hash from "./hash";
+export * from "./hash";
 export * as cookies from "./cookies";
 export * as jwt from "./jwt";